feat: move NewEventData into TimelineEvent
