+++ frontend/src/service/TempServerFacadeImpl.ts
@@ -2,11 +2,11 @@ import IServerFacade from "@/service/IServerFacade";
 import { mockContacts } from "@/utils/mockContacts";
 import { mockTimelineEvents } from "@/utils/mockTimelineEvents";
 import { ContactListQuery, ContactListQueryResult, NewContactData } from "@/service/ContactService";
-import { NewEventData, TimelineQuery, TimelineQueryResult } from "@/service/EventService";
+import { TimelineQuery, TimelineQueryResult } from "@/service/EventService";
 import { Category } from "@/model/Category";
 import { AuthToken } from "@/model/AuthToken";
 import { Contact } from "@/model/Contact";
-import { TimelineEvent } from "@/model/TimelineEvent";
+import { NewEventData, TimelineEvent } from "@/model/TimelineEvent";
 import { mockAuthToken } from "@/utils/mockAuthToken";
 
 const CATEGORY: Category = {
