+++ frontend/test/service/EventService.test.ts
@@ -2,8 +2,8 @@ import { beforeEach, describe, it } from "@jest/globals";
 import IServerFacade from "@/service/IServerFacade";
 import { instance, mock, verify } from "@typestrong/ts-mockito";
 import { AuthToken } from "@/model/AuthToken";
-import { TimelineEvent } from "@/model/TimelineEvent";
-import EventService, { NewEventData } from "@/service/EventService";
+import { NewEventData, TimelineEvent } from "@/model/TimelineEvent";
+import EventService from "@/service/EventService";
 
 const NEW_EVENT_DATA: NewEventData = {
   name: "EVENT-NAME",
